// components/ListingsMap.tsx
import { View, Text, StyleSheet } from "react-native";
import React from "react";
import MapView, { Marker, PROVIDER_GOOGLE } from "react-native-maps";
import { defaultStyles } from "@/constants/Styles";
import { ListingMapGeo } from "@/interfaces/listingMap";

interface Props {
  listingGeoData: any;
}

const INITIAL_REGION = {
  latitude: 51.165691,
  longitude: 10.451526,
  latitudeDelta: 9,
  longitudeDelta: 9,
};

const ListingsMap = ({ listingGeoData }: Props) => {
  return (
    <View style={defaultStyles.container}>
      <MapView
        style={StyleSheet.absoluteFill}
        provider={PROVIDER_GOOGLE}
        showsUserLocation
        showsMyLocationButton
        initialRegion={INITIAL_REGION}
      >
        {listingGeoData.features.map((item: ListingMapGeo) => (
          <Marker
            key={item.properties.id}
            coordinate={{
              latitude: +item.properties.latitude,
              longitude: +item.properties.longitude,
            }}
          />
        ))}
      </MapView>
    </View>
  );
};

const styles = StyleSheet.create({
  container: {
    flex: 1,
  },
});

export default ListingsMap;
